+++ demo_app/lib/screens/wordpress.dart
@@ -3,10 +3,9 @@ import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 import 'package:http/http.dart' as http;
 import 'package:url_launcher/url_launcher.dart';
-
-import '../html_widget.dart';
 
 class WordpressScreen extends StatelessWidget {
   final sites = {
