+++ packages/enhanced/README.md
@@ -62,7 +62,7 @@ class HelloWorldScreen extends StatelessWidget {
         body: SingleChildScrollView(
           child: Padding(
             padding: const EdgeInsets.all(8.0),
-            child: HtmlWidget(kHtml),
+            child: HtmlWidget(kHtml, webView: true),
           ),
         ),
       );
