+++ packages/enhanced/lib/src/widgets/web_view/fallback.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+
+import '../web_view.dart';
+
+class WebViewState extends State<WebView> {
+  @override
+  Widget build(BuildContext context) =>
+      AspectRatio(aspectRatio: widget.aspectRatio);
+}
